feat(controller): add user ownership guard

+++ server/src/controllers/portfolio/remove.ts
@@ -3,6 +3,7 @@ import provideRepos from '@server/trpc/provideRepos'
 import { portfolioRepository } from '@server/repositories/portfolioRepository'
 import { portfolioSchema } from '@server/entities/portfolio'
 import { TRPCError } from '@trpc/server'
+import { isUserPortfolioOwner } from '@server/utils/isUserPortfolioOwner'
 
 export default authenticatedProcedure
   .use(provideRepos({ portfolioRepository }))
@@ -12,6 +13,19 @@ export default authenticatedProcedure
     })
   )
   .mutation(async ({ input: portfolioData, ctx: { authUser, repos } }) => {
+    if (
+      !isUserPortfolioOwner(
+        portfolioData.id,
+        authUser.id,
+        repos.portfolioRepository
+      )
+    ) {
+      throw new TRPCError({
+        code: 'FORBIDDEN',
+        message: 'User do not have access to this portfolio.',
+      })
+    }
+
     const portfolioDeleted = await repos.portfolioRepository.remove(
       portfolioData.id,
       authUser.id
